Fix modal create account
