+++ package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@headlessui/react": "^1.7.17",
     "framer-motion": "10.16.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
